cafe-latte/solution/requirements.txt
galois==0.4.10
llvmlite==0.46.0
numba==0.64.0
numpy==2.4.3
pycryptodome==3.23.0
typing_extensions==4.15.0